[gradle][lsp] Fix errors in ProjectImportTest with Gradle 9.0.0-rc-3 by updating the Kotlin version up to 2.2.0
After updating Gradle up to 9.0.0-rc-3, these tests become failing for Kotlin version 1.9.10 because of this error:
java.lang.ClassNotFoundException: org.gradle.api.internal.HasConvention

The HasConvention class is not present in Gradle 9.0.0-rc-3, it was deleted. Probably, with the usage of Kotlin 1.9.10, it was accessed somehow. Updating the Kotlin version in build.gradle.kts resolved the issue.

With updating to the 2.2.0 Kotlin version, the content of workspace.json has changed:
- `kotlin-stdlib-common` disappeared there. Also, `kotlin-stdlib-jdk7` and `kotlin-stdlib-jdk8` for some tests. AFAIU, these are included in kotlin-stdlib as dependencies, so think it's okay.
- for `dependencies` test, `kotlin-stdlib-jdk7` and `kotlin-stdlib-jdk8` versions were downgraded from 1.9.10 to 1.8.21. It's not clear to me, why. But it seems okay because `kotlin-stdlib` depends on these libraries with versions 1.8.0 (lower than 2.2.0)

Also, updating to Kotlin 2.2.0 caused errors in MultiProjectKotlinDSL/build.gradle.kts: in this version, the usage of `kotlinOptions` is not supported. So, this commit replaces it with `compilerOptions`

GitOrigin-RevId: c2a899c9c0f4406788160fbc2894c02d7041e9c6