cleanup: remove unused dependencies on intellij.platform.inspect module (IJPL-200936)
This is needed to convert it to a content module and declare a runtime dependency only where needed.
A direct dependency is added to intellij.idea.community.main to run tests from the module. Also, some dependencies that came as transitive dependencies before now are added explicitly to language-server.project-import module.

GitOrigin-RevId: 406d8bb08840a6feed01e1513c1ed72fc060f090