[lsp] change diagnostic severity for ProblemHighlightType.GENERIC_ERROR_OR_WARNING
In IntelliJ, inspections with GENERIC_ERROR_OR_WARNING severity show up with yellow underline.
But for LSP, we were mapping it to `DiagnosticSeverity.Hint`.

As a result, in VSCode, these inspections were showing up as barely visible 3 dots at the start of the element.
And they also didn't appear in the Problems tab.

GitOrigin-RevId: 4a8f8bcf07edaddd39a713f954d265d7df7ca19e